+++ Services/Fsel.Course/Fsel.Course.Domain/Models/CommandModels/SectionTimeCodes/UpdateSectionTimeCodeCommandModel.cs
@@ -5,8 +5,8 @@ namespace Fsel.Course.Domain.Models.CommandModels.SectionTimeCodes
     public class UpdateSectionTimeCodeCommandModel
     {
         public string? Name { get; set; }
-        public long DisplayTime { get; set; }
-        public long ExecutionTime { get; set; }
+        public double DisplayTime { get; set; }
+        public double ExecutionTime { get; set; }
         public Guid SectionId { get; set; }
     }
 }
